tests/requirements.txt
# Core testing framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-xdist==3.3.1
pytest-mock==3.12.0
pytest-benchmark==4.0.0

# HTTP testing
httpx==0.25.2
requests==2.31.0
responses==0.24.0

# Database testing
pytest-postgresql==5.0.0
sqlalchemy==2.0.23
alembic==1.12.1

# Factory/Fixture generation
factory-boy==3.3.0
faker==20.1.0

# Performance testing
locust==2.17.0
pytest-benchmark==4.0.0

# API documentation testing
schemathesis==3.19.7

# Browser automation
playwright==1.40.0

# Test reporting
pytest-html==4.1.1
pytest-json-report==1.5.0
allure-pytest==2.13.2

# Scrapy testing
scrapy==2.11.0
itemadapter==0.8.0

# Async testing utilities
aioresponses==0.7.4
pytest-redis==3.0.2

# Code quality
flake8==6.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Security testing
bandit==1.7.5
safety==2.3.5

# Container testing
testcontainers==3.7.1
docker==6.1.3

# Monitoring and metrics
prometheus-client==0.19.0
structlog==23.2.0